# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/friends": {
            "get": {
                "tags": [
                    "Friend"
                ],
                "summary": "Delete post by id",
                "operationId": "73f1c8418ad5b2d4dd27bf3b82c6dcc9",
                "responses": {
                    "200": {
                        "description": "The resource was successfully retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add-friend": {
            "post": {
                "tags": [
                    "Friend"
                ],
                "summary": "Login to the account",
                "operationId": "9d8c83c5da53748e7c09612ebf0f08a2",
                "parameters": [
                    {
                        "name": "friend_id",
                        "in": "query",
                        "description": "Enter your friend_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends-request": {
            "get": {
                "tags": [
                    "Friend"
                ],
                "summary": "get list of friends requests",
                "operationId": "5525073f5e7062a1e174003ca1c38b33",
                "responses": {
                    "200": {
                        "description": "The resource was successfully retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/confirm-friend/{id}": {
            "put": {
                "tags": [
                    "Friend"
                ],
                "summary": "Confirm your friend",
                "operationId": "27e6907df47e0bfc3600de184a6247d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The resource was successfully retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-friend/{id}": {
            "delete": {
                "tags": [
                    "Friend"
                ],
                "summary": "Delete friend",
                "operationId": "b597b420001a06f2e19525fc5db8ce07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The resource was successfully retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/like": {
            "post": {
                "tags": [
                    "Like"
                ],
                "summary": "add comment to post",
                "operationId": "be176c487a370eb299781a9298e864b6",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "query",
                        "description": "Enter your id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/unlike/{id}": {
            "delete": {
                "tags": [
                    "Like"
                ],
                "summary": "Unlike a post",
                "operationId": "9caf84a69309b381e0b45ee4f0a09bc0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The resource was successfully retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    }
}